fix: cancel/task now returns the expected 'cancelled' task status (#158)
# Description

The current implementation of the cancel/task did not return the task in
'cancelled' status, because it was not consuming all the events on the
eventBus, but was taking the last published one.
This PR fixes this tck
[test](https://github.com/guglielmo-san/a2a-tck/blob/9893cfdb5f5ab661958c3edcfad43be806ad5626/tests/mandatory/protocol/test_tasks_cancel_method.py#L39)

Thank you for opening a Pull Request!
Before submitting your PR, there are a few things you can do to make
sure it goes smoothly:

- [x] Follow the [`CONTRIBUTING`
Guide](https://github.com/google-a2a/a2a-js/blob/main/CONTRIBUTING.md).
- [x] Make your Pull Request title in the
<https://www.conventionalcommits.org/> specification.
- Important Prefixes for
[release-please](https://github.com/googleapis/release-please):
- `fix:` which represents bug fixes, and correlates to a
[SemVer](https://semver.org/) patch.
- `feat:` represents a new feature, and correlates to a SemVer minor.
- `feat!:`, or `fix!:`, `refactor!:`, etc., which represent a breaking
change (indicated by the `!`) and will result in a SemVer major.
- [x] Ensure the tests and linter pass
- [x] Appropriate docs were updated (if necessary)

Fixes #157 🦕