fix: handle errors occurred in non-blocking sendMessage (#187)
# Description

This PR fixes the error of server crash in case of error in the event
loop for non blocking event, if the first reply to the client has been
already sent.
Two different error handling strategies are implemented based on the
origin of the error:

1. If the error originates from the pushNotification service, the loop
is not interrupted, and the error is simply printed on the console error
output
2. If the error is originating from the ExecutionEventQueue or from a
taskStore operation (resultManager interaction), the error is
interrupting the loop.
In the catch clause a TaskStatusUpdateEvent is generated, with
taskStatus equal to "failed", if an already existing Task has been
processed by the ResultManager. The new TaskStatusUpdateEvent is saved
in the taskStore. If a new error occurs within this operation, it is
simply printed on the console error output.

Fixes #186  🦕