feat: support authentication on server side (#195)
# Description

This PR integrates a support for authentication on the server side.
A new `User` interface is created to represent the Authenticated and
UnAuthenticated user in the context, with 2 required function:
1. isAuthenticathed()
2. userName()

The callback `authenticathedUserExtractor` provided to the
`A2AExpressApp` is responsible for extracting the user object from the
Request (if set up by the authentication middleware), and returning an
instance of the `User` interface. The returned instance will be added to
the ServerCallContext.
If no authentication middleware or no `authenticathedUserExtractor` is
provided to the Server, the user in the ServerCallContext will be of
type UnAuthenticated.

Fixes #194  🦕

Release-As: 0.3.6