ci: fix coverage comment workflow (#270)
# Description

Original approach from #262 isn't going to work for PRs from forks, as
such actions can't have write access to the upstream repo by default and
granting this access is not secure.

Following the approach with a dedicated
[`workflow_run`](https://docs.github.com/en/actions/reference/workflows-and-actions/events-that-trigger-workflows#workflow_run)
triggered workflow which is executed in a secure context.

Reference:
https://securitylab.github.com/resources/github-actions-preventing-pwn-requests/.

Tested in a fork: https://github.com/ishymko/a2a-js/pull/2.

Re #192